Refactor to decrease class and method lengths.
