+++ lib/rubocop.rb
@@ -100,6 +100,7 @@ require 'rubocop/cop/style/end_block'
 require 'rubocop/cop/style/end_of_line'
 require 'rubocop/cop/style/even_odd'
 require 'rubocop/cop/style/favor_join'
+require 'rubocop/cop/style/if_node'
 require 'rubocop/cop/style/favor_modifier'
 require 'rubocop/cop/style/favor_sprintf'
 require 'rubocop/cop/style/favor_unless_over_negated_if'
